+++ implementation/js/src/infrastructure/api.js
@@ -121,11 +121,10 @@
 
 				//see if we have app.viewSrcs set to load the View def dynamically
 				if(app.config && app.config.viewSrcs){
-					$.ajax({
-						url: _.compact([app.config.viewSrcs, type.toLowerCase(), app.nameToPath(name)]).join('/') + '.js',
-						dataType: 'script',
-						async: false
-					}).done(function(){
+					app.inject.js(
+						_.compact([app.config.viewSrcs, type.toLowerCase(), app.nameToPath(name)]).join('/') + '.js',
+						true //sync
+					).done(function(){
 						app.debug('View injected', name, 'from', app.viewSrcs);
 						Reusable = true;
 					}).fail(function(jqXHR, settings, e){
@@ -250,9 +249,16 @@
 
 		//-----------------dispatcher/observer/cache----------------
 		dispatcher: function(obj){ //+on/once, off; +listenTo/Once, stopListening; +trigger;
+			var dispatcher;
 			if(_.isPlainObject(obj))
-				return _.extend(obj, Backbone.Events);
-			return _.clone(Backbone.Events);
+				dispatcher = _.extend(obj, Backbone.Events);
+			else
+				dispatcher = _.clone(Backbone.Events);
+			dispatcher.dispose = function(){
+				this.off();
+				this.stopListening();
+			};
+			return dispatcher;
 		},
 
 		model: function(data){
@@ -303,6 +309,10 @@
 			window.location.reload();
 		},
 
+		//----------------fx animation---------------
+		nextFrame: window.requestAnimationFrame/*(next-step updater fn(t))*/,
+		cancelFrame: window.cancelAnimationFrame/*(frame id returned by nextFrame(fn))*/,
+
 		//----------------debug----------------------
 		debug: function(){
 			var fn = console.debug || console.log;
@@ -325,14 +335,16 @@
 	 * API summary
 	 */
 	app._apis = [
-		'model', 'collection',
-		'context - @alias:page', 'regional - @alias:area',
-		'view',
-		'widget', 'editor', 'editor.validator - @alias:editor.rule',
-		'remote',
-		'lock', 'unlock', 'available',
-		'download',
-		'create - @deprecated'
+		'dispatcher', 'model', 'collection',
+		'context - @alias:page', 'view', 'widget', 'editor', 'editor.validator - @alias:editor.rule', //view
+		'lock', 'unlock', 'available', //global action locks
+		'coop', 'navigate', 'reload', 'param', 'nextFrame', 'cancelFrame',
+		'remote', 'ws', 'download', //com
+		'extract', 'cookie', 'store', 'moment', 'uri', 'validator', //3rd-party lib short-cut
+		//@supportive
+		'debug', 'has', 'get', 'nameToPath', 'pathToName', 'inject.js', 'inject.tpl', 'inject.css',
+		//@deprecated
+		'create - @deprecated', 'regional - @deprecated'
 	];
 
 	/**